Convert channel metadata and links list to ts and pass link params
